refactor: :art: Moved output generation to a separate handler
